ebiten: single thread via build tag
