+++ audio/audio.go
@@ -37,8 +37,9 @@ import (
 	"fmt"
 	"io"
 	"runtime"
-	"sync"
 	"time"
+
+	"github.com/hajimehoshi/ebiten/v2/internal/sync"
 )
 
 const (
