+++ internal/buffered/command.go
@@ -15,8 +15,9 @@
 package buffered
 
 import (
-	"sync"
 	"sync/atomic"
+
+	"github.com/hajimehoshi/ebiten/v2/internal/sync"
 )
 
 var (
