+++ internal/clock/clock.go
@@ -16,8 +16,9 @@
 package clock
 
 import (
-	"sync"
 	"time"
+
+	"github.com/hajimehoshi/ebiten/v2/internal/sync"
 )
 
 var (
