+++ internal/glfw/glfw_notwindows.go
@@ -19,9 +19,9 @@ package glfw
 
 import (
 	"image"
-	"sync"
 
 	"github.com/go-gl/glfw/v3.3/glfw"
+	"github.com/hajimehoshi/ebiten/v2/internal/sync"
 )
 
 type windows map[*glfw.Window]*Window
