+++ internal/shareable/image.go
@@ -18,7 +18,6 @@ import (
 	"fmt"
 	"image/color"
 	"runtime"
-	"sync"
 
 	"github.com/hajimehoshi/ebiten/v2/internal/affine"
 	"github.com/hajimehoshi/ebiten/v2/internal/driver"
@@ -26,6 +25,7 @@ import (
 	"github.com/hajimehoshi/ebiten/v2/internal/hooks"
 	"github.com/hajimehoshi/ebiten/v2/internal/packing"
 	"github.com/hajimehoshi/ebiten/v2/internal/restorable"
+	"github.com/hajimehoshi/ebiten/v2/internal/sync"
 )
 
 const (
