+++ internal/thread/thread_noop.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// +build !ebitensinglethread
+// +build ebitensinglethread
 
 package thread
 
@@ -21,19 +21,13 @@ import (
 )
 
 // Thread represents an OS thread.
-type Thread struct {
-	funcs   chan func() error
-	results chan error
-}
+type Thread struct{}
 
 // New creates a new thread.
 //
 // It is assumed that the OS thread is fixed by runtime.LockOSThread when New is called.
 func New() *Thread {
-	return &Thread{
-		funcs:   make(chan func() error),
-		results: make(chan error),
-	}
+	return &Thread{}
 }
 
 // BreakLoop represents an termination of the loop.
@@ -42,16 +36,7 @@ var BreakLoop = errors.New("break loop")
 // Loop starts the thread loop until a posted function returns BreakLoop.
 //
 // Loop must be called on the thread.
-func (t *Thread) Loop() {
-	for f := range t.funcs {
-		err := f()
-		if err == BreakLoop {
-			t.results <- nil
-			return
-		}
-		t.results <- err
-	}
-}
+func (t *Thread) Loop() {}
 
 // Call calls f on the thread.
 //
@@ -61,6 +46,5 @@ func (t *Thread) Loop() {
 //
 // Call blocks if Loop is not called.
 func (t *Thread) Call(f func() error) error {
-	t.funcs <- f
-	return <-t.results
+	return f()
 }
