+++ internal/uidriver/glfw/input.go
@@ -19,11 +19,11 @@
 package glfw
 
 import (
-	"sync"
 	"unicode"
 
 	"github.com/hajimehoshi/ebiten/v2/internal/driver"
 	"github.com/hajimehoshi/ebiten/v2/internal/glfw"
+	"github.com/hajimehoshi/ebiten/v2/internal/sync"
 )
 
 type gamePad struct {
