+++ internal/uidriver/mobile/ui.go
@@ -19,7 +19,6 @@ package mobile
 import (
 	"fmt"
 	"runtime/debug"
-	"sync"
 	"sync/atomic"
 	"unicode"
 
@@ -37,6 +36,7 @@ import (
 	"github.com/hajimehoshi/ebiten/v2/internal/graphicsdriver/opengl"
 	"github.com/hajimehoshi/ebiten/v2/internal/hooks"
 	"github.com/hajimehoshi/ebiten/v2/internal/restorable"
+	"github.com/hajimehoshi/ebiten/v2/internal/sync"
 	"github.com/hajimehoshi/ebiten/v2/internal/thread"
 )
